requirements.txt

Flask>=3.0.0
python-dotenv>=1.0.1
rich==13.7.1
pynacl==1.5.0
base58==2.1.1
python-slugify==8.0.4
solana>=0.30,<0.35
solders>=0.21,<0.23
httpx==0.23.3
construct-typing==0.5.6
cachetools==4.2.4
prompt_toolkit==3.0.47
flask-swagger-ui>=4.11.1
flask-cors>=4.0.0

base58
cachetools
construct-typing
Flask
flask-cors
flask-swagger-ui
httpx
prompt_toolkit
pynacl
python-dotenv
python-slugify
rich
solana
solders
